refactor : 피드백 반영 - 자동차 이름 분리
