+++ src/main/java/racingcar/model/CarName.java
@@ -1,0 +1,17 @@
+package racingcar.model;
+
+public class CarName {
+
+    private final String name;
+
+    public CarName(String name) {
+        if (name.length() > 5) {
+            throw new IllegalArgumentException("자동차 이름은 5자를 초과할 수 없습니다.");
+        }
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
